1116_1 home

+++ libft/additional/ft_putstr_fd.c
@@ -1,19 +1,20 @@
 #include <unistd.h>
 #include "libft.h"
 
-static int	ft_strlen(char *str);
+static size_t	ft_strlen(char *str);
 
 void	ft_putstr_fd(char *s, int fd)
 {
 	write(fd, s, ft_strlen(s));
 }
 
-static int	ft_strlen(char *str)
+static size_t	ft_strlen(char *str)
 {
-	int	len;
+	size_t	len;
 
 	len = 0;
 	while (str[len])
 		len++;
 	return (len);
 }
+
